+++ docs/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - bump to osidb-bindings 4.15.0
 - update tools User Agent (aegis - https://github.com/RedHatProductSecurity/aegis) 
 - added some error handling for tools
+- add gemini safety settings
 
 ### Fixed